+++ benches/lurk.rs
@@ -1,7 +1,7 @@
 use criterion::{criterion_group, criterion_main, Criterion};
 use std::time::Duration;
 
-use loam::lurk::eval::build_lurk_toplevel_native;
+use loam::lurk::eval_direct::build_lurk_toplevel_native;
 
 fn toplevel(c: &mut Criterion) {
     c.bench_function("toplevel", |b| {
