+++ src/lurk/eval_direct.rs
